pyproject.toml
# pyproject.toml (V2.2 - Final Corrected Version with slowapi)

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-ai-core"
version = "0.1.0"
description = "An extensible, production-ready AI Agent platform built on FastAPI and Docker."
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "SolidFoundry", email = "ID+SolidFoundry@users.noreply.github.com" },
]
maintainers = [
    { name = "SolidFoundry", email = "ID+SolidFoundry@users.noreply.github.com" },
]
keywords = ["ai", "fastapi", "postgresql", "openai", "llm", "agent", "python", "saas"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: FastAPI",
]
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.2.0",
    "openai>=1.16.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "python-json-logger>=2.0.0",
    "httpx[http2]>=0.27.0",
    "slowapi>=0.1.9",
    "cachetools>=5.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=5.0.0",
    "black>=24.4.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
]
test = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=5.0.0",
]

[project.urls]
"Homepage" = "https://github.com/SolidFoundry/py-ai-core"
"Repository" = "https://github.com/SolidFoundry/py-ai-core"
"Bug Tracker" = "https://github.com/SolidFoundry/py-ai-core/issues"
"Discussions" = "https://github.com/SolidFoundry/py-ai-core/discussions"

[tool.setuptools]
packages = ["py_ai_core"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["py_ai_core"]